Move vocab learn infra into chat tools
