+++ algorithm/lazymind/chat/engine/tools/infra/vocab_registry.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import threading
 
-from lazymind.review.vocab import VocabManager
-from lazymind.review.service.db import fetch_vocab_for_user_id
+from .vocab_db import fetch_vocab_for_user_id
+from .vocab_manager import VocabManager
 
 _registry: dict[str, VocabManager] = {}
 _registry_lock = threading.Lock()
@@ -17,9 +17,11 @@ def get_vocab_manager(user_id: str = '') -> VocabManager:
                     user_id=user_id,
                     data_source=lambda: fetch_vocab_for_user_id(user_id),
                 )
-    return _registry[user_id]
+    manager = _registry[user_id]
+    manager.reload()
+    return manager
 
 
-def clear_registry() -> None:
+def clear_vocab_registry() -> None:
     with _registry_lock:
         _registry.clear()
