+++ algorithm/lazymind/review/api/__init__.py
@@ -1,7 +1,3 @@
 from __future__ import annotations
 
-from . import vocab_routes
-
-__all__ = [
-    'vocab_routes',
-]
+__all__: list[str] = []
