+++ algorithm/lazymind/review/service/__init__.py
@@ -1,17 +1,3 @@
 """Review service modules."""
 
-from .evolution import (
-    VocabEvolutionService,
-    apply_vocab_evolution_actions,
-    get_vocab_evolution_service,
-    resolve_word_group_apply_url,
-    run_vocab_evolution,
-)
-
-__all__ = [
-    'VocabEvolutionService',
-    'apply_vocab_evolution_actions',
-    'get_vocab_evolution_service',
-    'resolve_word_group_apply_url',
-    'run_vocab_evolution',
-]
+__all__: list[str] = []
