--- algorithm/lazymind/review/vocab/__init__.py
@@ -1,31 +1,0 @@
-from .evolution import (
-    ActionPlanningModule,
-    ChatHistoryRecord,
-    HistoryChunker,
-    HistoryCollector,
-    LAZYLLM_CONTEXT_CREATE_USER_ATTR,
-    SynonymCandidate,
-    SynonymExtractionModule,
-    VocabEvolutionRequest,
-    get_ppl_vocab_evolution,
-    json_dump_list,
-    norm_text,
-    summarize_action_for_log,
-)
-from .vocab_manager import VocabManager
-
-__all__ = [
-    'ActionPlanningModule',
-    'ChatHistoryRecord',
-    'HistoryChunker',
-    'HistoryCollector',
-    'LAZYLLM_CONTEXT_CREATE_USER_ATTR',
-    'SynonymCandidate',
-    'SynonymExtractionModule',
-    'VocabEvolutionRequest',
-    'VocabManager',
-    'get_ppl_vocab_evolution',
-    'json_dump_list',
-    'norm_text',
-    'summarize_action_for_log',
-]
